% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFprofileR_api.R
\name{print.nmf_profile}
\alias{print.nmf_profile}
\title{Print an NMFprofileR result}
\usage{
\method{print}{nmf_profile}(x, ...)
}
\arguments{
\item{x}{An `nmf_profile` object returned (invisibly) by [NMFprofileR()].}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print an NMFprofileR result
}
