% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_expression_data}
\alias{example_expression_data}
\title{Example Expression Data for NMFprofileR}
\format{
A matrix with 100 rows and 20 columns.
}
\usage{
example_expression_data
}
\description{
Example GBM Mesenchymal-like subtype expression data from PanCancerAtlas
}
\keyword{datasets}
